@@ -4,6 +4,14 @@ import { CategoryComponent } from './category/category.component';
 import { ArticleComponent } from './article/article.component';
 import { CartComponent } from './cart/cart.component';
 import { pagesRoutingModule } from './pages-routing.module';
+import { DataTablesModule } from 'angular-datatables';
+import { AddCategorieComponent } from './add-categorie/add-categorie.component';
+import { RouterModule } from '@angular/router';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { AddArticuloComponent } from './add-articulo/add-articulo.component';
+import { EditArticuloComponent } from './edit-articulo/edit-articulo.component';
+
+
 
 
 
@@ -12,10 +20,17 @@ import { pagesRoutingModule } from './pages-routing.module';
     CategoryComponent,
     ArticleComponent,
     CartComponent,
-    CategoryComponent
+    CategoryComponent,
+    AddCategorieComponent,
+    AddArticuloComponent,
+    EditArticuloComponent
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    DataTablesModule,
+    RouterModule,
+    FormsModule,
+    ReactiveFormsModule
   ],exports:[
     pagesRoutingModule
   ]
