@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Route, Router } from '@angular/router';
 
 @Component({
   selector: 'app-navbar',
@@ -7,9 +8,23 @@ import { Component, OnInit } from '@angular/core';
 })
 export class NavbarComponent implements OnInit {
 
-  constructor() { }
+  constructor(private activeRouter:Router) { }
 
   ngOnInit(): void {
   }
+
+  logout(){
+      localStorage.removeItem('token'); 
+      this.activeRouter.navigate(['/home'])
+  }
+
+  isLoggedIn(){
+    if(localStorage.getItem('token')){
+        return true
+    }
+    else{
+        return false
+    }
+}
 
 }
